Handle yelp API error
Be aware that Yelp API calls might return errors. Extract body from the call in
a synchronized manner. Any errors will be handled directly. The privious version
of the YelpResturantFetcher tried to extract the response body at a later stage,
which postponed the error handling. I.e. the exceptions was thrown when a user
asked for a resturant since it wasn't until then that the response body was
extracted. The thrown exception was then handled by Spring Boot's default
exception handler.

The YelpResturantFetcher might return null if a resturant object hasn't been set,
the resturant service has to be able to handle that. It will currently throw a
NullPointerException.